code/requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
_openmp_mutex=5.1=1_gnu
addict=2.4.0=pypi_0
aiohttp=3.9.3=pypi_0
aiosignal=1.3.1=pypi_0
albumentations=1.4.2=pypi_0
appdirs=1.4.4=pypi_0
asttokens=2.4.1=pypi_0
async-timeout=4.0.3=pypi_0
attrs=23.2.0=pypi_0
backcall=0.2.0=pypi_0
blinker=1.7.0=pypi_0
ca-certificates=2024.3.11=h06a4308_0
certifi=2024.2.2=pypi_0
cffi=1.16.0=pypi_0
charset-normalizer=3.3.2=pypi_0
click=8.1.7=pypi_0
comm=0.2.2=pypi_0
configargparse=1.7=pypi_0
contourpy=1.1.1=pypi_0
cycler=0.12.1=pypi_0
dash=2.16.1=pypi_0
dash-core-components=2.0.0=pypi_0
dash-html-components=2.0.0=pypi_0
dash-table=5.0.0=pypi_0
decorator=5.1.1=pypi_0
docker-pycreds=0.4.0=pypi_0
efficientnet-pytorch=0.7.1=pypi_0
executing=2.0.1=pypi_0
fastjsonschema=2.19.1=pypi_0
filelock=3.13.3=pypi_0
flask=3.0.2=pypi_0
fonttools=4.50.0=pypi_0
frozenlist=1.4.1=pypi_0
fsspec=2024.3.1=pypi_0
gitdb=4.0.11=pypi_0
gitpython=3.1.42=pypi_0
h5py=3.10.0=pypi_0
huggingface-hub=0.22.1=pypi_0
idna=3.6=pypi_0
iglovikov-helper-functions=0.0.53=pypi_0
imagecorruptions=1.1.2=pypi_0
imageio=2.34.0=pypi_0
importlib-metadata=7.1.0=pypi_0
importlib-resources=6.4.0=pypi_0
ipython=8.12.3=pypi_0
ipywidgets=8.1.2=pypi_0
itsdangerous=2.1.2=pypi_0
jedi=0.19.1=pypi_0
jinja2=3.1.3=pypi_0
joblib=1.3.2=pypi_0
jpeg4py=0.1.4=pypi_0
jsonschema=4.21.1=pypi_0
jsonschema-specifications=2023.12.1=pypi_0
jupyter-core=5.7.2=pypi_0
jupyterlab-widgets=3.0.10=pypi_0
kiwisolver=1.4.5=pypi_0
lazy-loader=0.3=pypi_0
ld_impl_linux-64=2.38=h1181459_1
libffi=3.4.4=h6a678d5_0
libgcc-ng=11.2.0=h1234567_1
libgomp=11.2.0=h1234567_1
libstdcxx-ng=11.2.0=h1234567_1
lightning-utilities=0.11.2=pypi_0
markupsafe=2.1.5=pypi_0
matplotlib=3.7.5=pypi_0
matplotlib-inline=0.1.6=pypi_0
mpmath=1.3.0=pypi_0
multidict=6.0.5=pypi_0
munch=4.0.0=pypi_0
nbformat=5.10.3=pypi_0
ncurses=6.4=h6a678d5_0
nest-asyncio=1.6.0=pypi_0
networkx=3.1=pypi_0
numpy=1.24.4=pypi_0
nvidia-cublas-cu12=12.1.3.1=pypi_0
nvidia-cuda-cupti-cu12=12.1.105=pypi_0
nvidia-cuda-nvrtc-cu12=12.1.105=pypi_0
nvidia-cuda-runtime-cu12=12.1.105=pypi_0
nvidia-cudnn-cu12=8.9.2.26=pypi_0
nvidia-cufft-cu12=11.0.2.54=pypi_0
nvidia-curand-cu12=10.3.2.106=pypi_0
nvidia-cusolver-cu12=11.4.5.107=pypi_0
nvidia-cusparse-cu12=12.1.0.106=pypi_0
nvidia-nccl-cu12=2.19.3=pypi_0
nvidia-nvjitlink-cu12=12.4.99=pypi_0
nvidia-nvtx-cu12=12.1.105=pypi_0
open3d=0.18.0=pypi_0
opencv-python=4.9.0.80=pypi_0
opencv-python-headless=4.9.0.80=pypi_0
openssl=3.0.13=h7f8727e_0
packaging=24.0=pypi_0
panda=0.3.1=pypi_0
pandas=2.0.3=pypi_0
parso=0.8.3=pypi_0
pexpect=4.9.0=pypi_0
pickleshare=0.7.5=pypi_0
pillow=10.2.0=pypi_0
pip=23.3.1=py38h06a4308_0
pkgutil-resolve-name=1.3.10=pypi_0
platformdirs=4.2.0=pypi_0
plotly=5.20.0=pypi_0
plyfile=1.0.3=pypi_0
pooch=1.8.1=pypi_0
pretrainedmodels=0.7.4=pypi_0
progressbar=2.5=pypi_0
progressbar2=4.4.2=pypi_0
prompt-toolkit=3.0.43=pypi_0
protobuf=4.25.3=pypi_0
psutil=5.9.8=pypi_0
ptyprocess=0.7.0=pypi_0
pure-eval=0.2.2=pypi_0
pycparser=2.21=pypi_0
pygments=2.17.2=pypi_0
pylab-sdk=1.7.2=pypi_0
pyntcloud=0.3.1=pypi_0
pyparsing=3.1.2=pypi_0
pyquaternion=0.9.9=pypi_0
python=3.8.19=h955ad1f_0
python-dateutil=2.9.0.post0=pypi_0
python-utils=3.8.2=pypi_0
pytorch-lightning=2.2.1=pypi_0
pytorch-toolbelt=0.6.3=pypi_0
pytz=2024.1=pypi_0
pyvista=0.43.4=pypi_0
pywavelets=1.4.1=pypi_0
pyyaml=6.0.1=pypi_0
readline=8.2=h5eee18b_0
referencing=0.34.0=pypi_0
requests=2.31.0=pypi_0
retrying=1.3.4=pypi_0
rpds-py=0.18.0=pypi_0
safetensors=0.4.2=pypi_0
scikit-image=0.21.0=pypi_0
scikit-learn=1.3.2=pypi_0
scipy=1.10.1=pypi_0
scooby=0.9.2=pypi_0
seaborn=0.13.2=pypi_0
segmentation-models-pytorch=0.3.3=pypi_0
sentry-sdk=1.43.0=pypi_0
setproctitle=1.3.3=pypi_0
setuptools=68.2.2=py38h06a4308_0
six=1.16.0=pypi_0
smmap=5.0.1=pypi_0
sqlite=3.41.2=h5eee18b_0
stack-data=0.6.3=pypi_0
sympy=1.12=pypi_0
tenacity=8.2.3=pypi_0
threadpoolctl=3.4.0=pypi_0
tifffile=2023.7.10=pypi_0
timm=0.9.2=pypi_0
tk=8.6.12=h1ccaba5_0
torch=2.2.2=pypi_0
torchmetrics=1.3.2=pypi_0
torchvision=0.17.2=pypi_0
tqdm=4.66.2=pypi_0
traitlets=5.14.2=pypi_0
triton=2.2.0=pypi_0
typing-extensions=4.10.0=pypi_0
tzdata=2024.1=pypi_0
urllib3=2.2.1=pypi_0
vtk=9.3.0=pypi_0
wandb=0.16.5=pypi_0
wcwidth=0.2.13=pypi_0
werkzeug=3.0.1=pypi_0
wheel=0.41.2=py38h06a4308_0
widgetsnbextension=4.0.10=pypi_0
xz=5.4.6=h5eee18b_0
yarl=1.9.4=pypi_0
zipp=3.18.1=pypi_0
zlib=1.2.13=h5eee18b_0
